- Rewrote the enemy ai (still doesnt work) - Added win and gameover menu - Player can now die - Added cheese collectable - Added spawning of cheese and enemies

+++ HKU_TurnBasedRoguelike/Assets/Scripts/Enemies/Enemy.cs
@@ -14,7 +14,7 @@ public class Enemy : ScriptableObject
     public int movementRange = 1;
     public float movementSpeed = 8;
     [Header("Combat")]
-    public float attackRange = 1;
+    public int attackRange = 1;
     public float minDamage = 60;
     public float maxDamage = 120;
     public float damageMultiplier = 1f;
